pyproject.toml
[project]
name = "snapchat-all-memories-downloader"
version = "1.0.0"
description = "Download all your Snapchat memories with timestamps, geolocation, and overlays"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Contributors", email = ""}
]
keywords = ["snapchat", "memories", "downloader", "ocr", "geolocation"]
classifiers = [
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Video",
    "Topic :: Utilities"
]
dependencies = [
    "httpx>=0.28.1",
    "piexif>=1.1.3",
    "Pillow>=12.0.0",
    "pydantic>=2.12.3",
    "tqdm>=4.67.1",
    "timezonefinder",
    "pytz",
    "easyocr",
    "numpy"
]
